mh swing rate function
V2.14 Update:
Added a menu that allows player to change the main hand swing rate in duel wield mode.
As we know when player duel wielding same type weapon , we get same log of weapon attack message. So it's unable to tell which from mh , which from oh. But the rate should be a static number. It's up to player to test. my test result is in 100% haste buff , MH swing rate = 56.9%. So that we can estimate an agro amount we have at the moment.
BTW It seems that haste matters in this rate. I guess class matters too due to different skill cap of DW. again ,this is up to test by players. I can adapt the code with more information gathered