v2.17
V2.17 update:
-optimize some logic for clicky aggro. when you zone out, cycle back or /t clearam in game , your accumlated clicky aggro will be cleared totally.

reason: bc the eq log file won't tell which target your spell land on (it only tells you that your xyz spell get resisted or abc get hit by some effect not sure if its' from you.) , so wha t Aggro Metter does is it will first accumulate the spell you casted into a individual variable ,named spell aggro then it will add this spell aggro to the next target when you swing on. This is assuming that your next swing will be on same target as your previous cast. If you changed your target with clicky aggro on old one and without hitting it at all. this would confuse the tool. So a in game quick corrective option is added in this patch ,that is zone it or type /t clearam in game or cycle back.