CCHPM V3.01
Raid Attendance Log Taker function is available.
This patch added a module that catches the !KLOG and corresponded logs after it. After got whole log and tells, it converts these logs lines into a formatted text that fit for our discord bot. So it will can add/swap characters automatically. This convert  based on an AI module named scikit-learn(open sourced too) , it requires some training based on our tells. we ll need report bugs here and accumulate the format people use for tells for log and training it more times as needed. Also based on our roster which is saved in the Raiders.txt. So update this raiders.txt periodically to make the result more accurate.